Add files via upload
This is a game I am lead developer and programmer on for my GMAP 260 class Spring of 2017.  This game takes a dashing mechanic and pits two players up against each other.  We have four weeks to develop this game, and I have three teammates working on art and documentation.  The game will be bug themed and feature a second mechanic that determines the winner of the battle after the two player crash into each other